+++ HISTORY.md
@@ -1,7 +1,17 @@
 # Historical logs
 
+## Version 0.2.1
+
+- Release date: Dec 02, 2020
+- Patched to use mctestgo v0.5.0
+
+## Version 0.2.0
+
+- Release date: Dec 02, 2020
+- Refactored cache functions and use mctestgo v0.4.0
+
 ## Version 0.1.0
 
-- Release date: TBD
+- Release date: Dec 01, 2020
 - Simple and Hash In-Memory Cache
 - See the test folder for different test cases / scenarios and usage